:refactor: Split base classes from LoadScene.cs, UnloadScene.cs
